YT Transcriber Summarizer/requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
watchdog